+++ libft/ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: stherkil <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/11 15:04:50 by stherkil          #+#    #+#             */
-/*   Updated: 2019/01/25 15:41:40 by stherkil         ###   ########.fr       */
+/*   Updated: 2019/08/31 15:13:47 by stherkil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		ft_putnbr(int n)
+void		ft_putnbr(long long n)
 {
 	ft_putnbr_fd(n, 1);
 }
