+++ libft/ft_putoct.c
@@ -6,15 +6,15 @@
 /*   By: stherkil <stherkil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/11 15:04:50 by stherkil          #+#    #+#             */
-/*   Updated: 2019/08/24 15:18:57 by stherkil         ###   ########.fr       */
+/*   Updated: 2019/09/01 20:53:57 by stherkil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_putoct(int n)
+intmax_t		ft_putoct(intmax_t n)
 {
-	int result;
+	intmax_t result;
 
 	result = 0;
 	if (n < 8)
